% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptlmapper.R
\name{build_kd_matrix}
\alias{build_kd_matrix}
\title{A Function That Builds Kantorivitch Distance Matrix}
\usage{
build_kd_matrix(hists)
}
\arguments{
\item{hists}{A list of histograms.}
}
\description{
This function builds Kantorivitch distance matrix from a list of histograms with all the same `breaks` value.
}

