% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptlmapper.R
\name{build_mmoments_matrix}
\alias{build_mmoments_matrix}
\title{A Function That Builds Multivariate Moment Matrix.}
\usage{
build_mmoments_matrix(hists, nb_moments = 4)
}
\arguments{
\item{hists}{A list of histograms.}

\item{nb_moments}{An integer specifying the number of moments to compute.}
}
\description{
This function builds multivariate moment matrix from a list of histograms.
}

