% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptlmapper.R
\name{build_pheno_hists}
\alias{build_pheno_hists}
\title{A Function That Builds Histograms From Single Cell Data}
\usage{
build_pheno_hists(cells, bin_width = NULL, nb_bin = 100)
}
\arguments{
\item{cells}{A list of vector of integer (single cell values).}

\item{bin_width}{An integer specifying the width of the bin to use.}

\item{nb_bin}{An integer specifying the number of to use.}
}
\description{
This function builds histograms form single cell data.
}

