% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptlmapper.R
\name{check_marker_allele_prop}
\alias{check_marker_allele_prop}
\title{A Function That Checks Proportion of Allele for each Marker.}
\usage{
check_marker_allele_prop(genodata, bckg, min_prop = 0.1)
}
\arguments{
\item{genodata}{A matrix describing the genotype of individuals.}

\item{bckg}{Vector of character describing the individuals as they are describe in genodata.}

\item{min_prop}{A numeric specifying the minimal proportion of each parental allele under which a marker is discard from the analysis.}
}
\description{
This function checks the proportion of allele for each marker. 
It discard marker that does not repsect the minimal proportion for an allele.
}

