% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptlmapper.R
\name{extract_axis_info}
\alias{extract_axis_info}
\title{A Function That Extracts Axis Information from a 'ptl_mapping' Data Structure}
\usage{
extract_axis_info(ptl_mapping_result, delta = 0.6)
}
\arguments{
\item{ptl_mapping_result}{A `ptl_mapping` data structure.}

\item{delta}{A numeric proportional to the space between chromosomes on the x axis.}
}
\description{
This function extracts axis information from a 'ptl_mapping' data structure.
}

