% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptlmapper.R
\name{get_best_markers_rqtl}
\alias{get_best_markers_rqtl}
\title{A Function That Returns Best Markers Obtained in the Classical QTL Search.}
\usage{
get_best_markers_rqtl(ptl_mapping_result, which_pheno)
}
\arguments{
\item{ptl_mapping_result}{...}

\item{which_pheno}{...}
}
\description{
This function returns best markers obtained in the classical QTL search.
}

