% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptlmapper.R
\name{get_wilks_score}
\alias{get_wilks_score}
\title{A Function That Compute the Wilks Score}
\usage{
get_wilks_score(data, all)
}
\arguments{
\item{data}{Multivariate coordinates of individuals.}

\item{all}{A vector of factors that describes groups of individuals.}
}
\description{
This function compute the Wilks score. The score formula is extracted from source code of the candisc package.
}

