% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptlmapper.R
\name{kantorovich}
\alias{kantorovich}
\title{A Function That Computes Kantorovich Distance.}
\usage{
kantorovich(x, y, nbreaks = 1000, lims = NULL, h1 = NULL, h2 = NULL)
}
\arguments{
\item{x}{...}

\item{y}{...}

\item{nbreaks}{...}

\item{lims}{...}

\item{h1}{...}

\item{h2}{...}
}
\description{
This function computes Kantorovich distance.
}

