% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptlmapper.R
\name{marker2col}
\alias{marker2col}
\title{A Function That Converts Marker Names Into Color Index}
\usage{
marker2col(ptl_mapping_result, marker_name, method = "kanto")
}
\arguments{
\item{ptl_mapping_result}{An outputs of the `ptl_mapping` function.}

\item{marker_name}{A character string specifying on which marker we are focused.}

\item{method}{A character string in c("kanto", "mmoments") that specify the method to use to characterize phenotypic distribution.}
}
\description{
This function converts marker names Into color index.
}

