% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptlmapper.R
\name{mn_2_col}
\alias{mn_2_col}
\title{A Function That Converts Marker Names Into Color Index.}
\usage{
mn_2_col(ptl_mapping_result, marker_names)
}
\arguments{
\item{ptl_mapping_result}{...}

\item{marker_names}{...}
}
\description{
This function converts marker names Into color index.
}

