% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptlmapper.R
\name{plot_can}
\alias{plot_can}
\title{A Function That Plots the Linear Disciminant Analysis.}
\usage{
plot_can(ptl_mapping_result, marker_names, main = "", type = "raw",
  col = NULL)
}
\arguments{
\item{ptl_mapping_result}{...}

\item{marker_names}{...}

\item{main}{...}

\item{type}{...}

\item{col}{...}
}
\description{
This function plots the linear disciminant analysis.
}

