% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptlmapper.R
\name{plot_dist}
\alias{plot_dist}
\title{A Function Distribution of the Phenotypes}
\usage{
plot_dist(ptl_mapping_result, col = NULL, main = "", xlim = NULL,
  ylim = NULL, only_n_first = 0, KANTO_DENSITY = TRUE,
  MEAN_DENSITY = FALSE, ...)
}
\arguments{
\item{ptl_mapping_result}{...}

\item{col}{...}

\item{main}{...}

\item{xlim}{...}

\item{ylim}{...}

\item{only_n_first}{...}

\item{KANTO_DENSITY}{...}

\item{MEAN_DENSITY}{...}

\item{...}{...}
}
\description{
This function plots distribution of the phenotypes.
}

