% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptlmapper.R
\name{plot_empirical_test}
\alias{plot_empirical_test}
\title{A Function That Plots Empirical P-Values}
\usage{
plot_empirical_test(ptl_mapping_result, main = "", errs = c(0.05, 0.01,
  0.005))
}
\arguments{
\item{ptl_mapping_result}{...}

\item{main}{...}

\item{errs}{...}
}
\description{
This function plots empirical p-values.
}

