% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptlmapper.R
\name{plot_mds}
\alias{plot_mds}
\title{A Function That Plots MDS.}
\usage{
plot_mds(ptl_mapping_result, main = "", type = "raw", col = NULL,
  mds = NULL, ...)
}
\arguments{
\item{ptl_mapping_result}{...}

\item{main}{...}

\item{type}{...}

\item{col}{...}

\item{mds}{...}
}
\description{
This function plots MDS.
}

