% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptlmapper.R
\name{plot_noise}
\alias{plot_noise}
\title{A Function That Plots Noise}
\usage{
plot_noise(ptl_mapping_result, main = "", col = NULL)
}
\arguments{
\item{ptl_mapping_result}{...}

\item{main}{...}

\item{col}{...}
}
\description{
This function plots noise.
}

