% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptlmapper.R
\name{plot_orth_trans}
\alias{plot_orth_trans}
\title{A Function That Plots Orthogonal Transformation}
\usage{
plot_orth_trans(ptl_mapping_result, method = "kanto", main = "",
  col = NULL, orth_trans = NULL, ...)
}
\arguments{
\item{ptl_mapping_result}{An outputs of the `ptl_mapping` function.}

\item{method}{A character string in c("kanto", "mmoments") that specify the method to use to characterize phenotypic distribution.}

\item{main}{A character string specifying additionnal title informatoion.}

\item{col}{Forcing col value of the plot.}

\item{orth_trans}{Provide your own orthogonal transformation (experimental!).}

\item{...}{Args forward to `plot` function.}
}
\description{
This function plots orthogonal transformation.
}

