% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptlmapper.R
\name{plot_wilks_k}
\alias{plot_wilks_k}
\title{A Function That Plots Encapsulated Wilks Scores for Kantoritch Approach}
\usage{
plot_wilks_k(...)
}
\arguments{
\item{...}{Args forward to `plot_wilks` function.}
}
\description{
This function plots encapsulated Wilks scores for Kantoritch approach.
}

