% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptlmapper.R
\name{plot_wilks_mm}
\alias{plot_wilks_mm}
\title{A Function That Plots Encapsulated Wilks Scores for MultiVariate Moments Approach.}
\usage{
plot_wilks_mm(ptl_mapping_result, main = "", ...)
}
\arguments{
\item{ptl_mapping_result}{...}

\item{main}{...}
}
\description{
This function plots encapsulated Wilks scores for multiVariate moments approach.
}

