% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptlmapper.R
\name{preprocess_genodata_rqtl}
\alias{preprocess_genodata_rqtl}
\title{A Function That Preprocesses Genotypes of Individuals for R Package QTL.}
\usage{
preprocess_genodata_rqtl(genodata, bckg)
}
\arguments{
\item{genodata}{A matrix describing the genotype of individuals.}

\item{bckg}{Vector of character describing the individuals as they are describe in genodata.}
}
\description{
This function preprocesses genotypes of individuals for R package QTL. 
In geneodata, "2" stands for undetermined and is replaced by NA.
}

