% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptlmapper.R
\name{preprocess_phenodata_rqtl}
\alias{preprocess_phenodata_rqtl}
\title{A Function That Preprocesses Phenotypes of Individuals for R Package QTL}
\usage{
preprocess_phenodata_rqtl(pheno_hists, kanto_analysis = NULL,
  mmoments_analysis = NULL)
}
\arguments{
\item{pheno_hists}{A list of object containing mean and var attribute. Typically outpout of the `build_pheno_hists` function.}

\item{kanto_analysis}{Output of `ptl_scan` function used with the method "kanto".}

\item{mmoments_analysis}{Output of `ptl_scan` function used with the method "mmoments".}
}
\description{
This function preprocesses phenotypes of individuals for R Package QTL.
}

