% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptlmapper.R
\name{ptl_scan}
\alias{ptl_scan}
\title{A Function That Scans Genome to Detect PTLs.}
\usage{
ptl_scan(pheno_matrix, genodata_ptl, nb_perm, nb_dim = 0, min_prop = 0.1,
  SHOW_PERM_PROG = TRUE, perm_prog_freq = 5, method = "kanto")
}
\arguments{
\item{genodata_ptl}{The preprocessed genodata. Typically output of `preprocess_genodata` function.}

\item{nb_perm}{An integer that specifies the number of permuation to do.}

\item{nb_dim}{An integer that specifies the number of dimension of the MDS space to explore.}

\item{min_prop}{A numeric specifying the minimal proportion of each parental allele under which a marker is discard from the analysis.}

\item{method}{A character string in c("kanto", "mmoments") that specify the method to use to characterize phenotypic distribution. Default is our favourite: "kanto"!}

\item{kd_matrix}{The Kantorivitch distance matrix.}

\item{SHOW_PERM_PROG=TRUE}{A boolean specifying if permutation progression need to to report on console.}

\item{perm_prog_freq=5}{An integer specifying the frequency of the permution progression reporting.}
}
\description{
This function scans genome to detect PTLs. 
It determines with the Kaiser criterion how many dimensions of the MDS are relevants for canonical analysis (`eig_to_scan`).
If nb_dim=0, a z-core criterion id used to determine with how many dimensions the canonical analysis gives the best result (to fixe 2 <= nb_dim <= eig_to_scan).
If nb_dim=NULL, `nb_dim` is automatically set to eig_to_scan.
If nb_dim>0 the canonical analysis is performed used provided `nb_dim`. In this case, keep attention to the fact that a high value of `nb_dim` introduce a lot of degree of freedom in your canonical analysis. The number of dimension `nb_dim` needs to be __substantially__ smaller than the size of your population. 
The permutations batch uses `ptl_scan` function with the `nb_dim` value as the one provided by the initial ptl_scan call.
}

