% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptlmapper.R
\name{rqtl_launch}
\alias{rqtl_launch}
\title{A Function That Scans Genome to Detect QTLs (based on R package `qtl`)}
\usage{
rqtl_launch(genodata_ptl, pheno_hists, kanto_analysis = NULL,
  mmoments_analysis = NULL, nb_perm = 1000, errs = 0.05,
  SKEWNESS_AND_KURTOSIS = FALSE)
}
\arguments{
\item{genodata_ptl}{The preprocessed genodata. Typically output of `preprocess_genodata` function.}

\item{pheno_hists}{A list of object containing mean and var attribute. Typically outpout of the `build_pheno_hists` function.}

\item{kanto_analysis}{Output of `ptl_scan` function used with the method "kanto".}

\item{mmoments_analysis}{Output of `ptl_scan` function used with the method "mmoments".}

\item{nb_perm}{An integer that specifies the number of permuation to do.}

\item{errs}{A vector of integer (error) that will be used to compute threshold from the permutation test.}

\item{SKEWNESS_AND_KURTOSIS}{A boolean that specifies if skewness and kurtosis QTl need to be scanned.}
}
\description{
This function embed R package `qtl`. It scans genome to detect QTLs using `ptlmapper` data structires.
}

