% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptlmapper.R
\name{zscore}
\alias{zscore}
\title{A Function That Computes Z-Score.}
\usage{
zscore(dWs, p = 0.95)
}
\arguments{
\item{dWs}{A vector of numeric.}

\item{p}{A numeric that specifies the quantile considered a noise.}
}
\value{
the z-score
}
\description{
This function returns z-score of serie.
}

