% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptlmapper.R
\name{get_best_markers_rptl}
\alias{get_best_markers_rptl}
\title{A Function That Returns Best Markers Obtained in the PTL Search.}
\usage{
get_best_markers_rptl(ptl_mapping_result, type = "raw", chr = NULL)
}
\arguments{
\item{ptl_mapping_result}{...}

\item{type}{...}

\item{chr}{...}
}
\description{
This function returns best markers obtained in the PTL search.
}

