% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptlmapper.R
\name{plot_rqtl}
\alias{plot_rqtl}
\title{A Function That Plots Encapsulated RQTL Outputs.}
\usage{
plot_rqtl(ptl_mapping_result, main = "", ylim = NULL, which_pheno = NULL)
}
\arguments{
\item{ptl_mapping_result}{...}

\item{main}{...}

\item{ylim}{...}

\item{which_pheno}{...}
}
\description{
This function plots encapsulated RQTL outputs.
}

