% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptlmapper.R
\name{plot_wilks}
\alias{plot_wilks}
\title{A Function That Plots Encapsulated Wilks Scores.}
\usage{
plot_wilks(ptl_mapping_result, main = "", type = "raw", pa = NULL,
  errs = c(0.05, 0.01, 0.005), ylim = NULL, y_thres = NULL)
}
\arguments{
\item{ptl_mapping_result}{...}

\item{main}{...}

\item{type}{...}

\item{pa}{...}

\item{errs}{...}

\item{ylim}{...}

\item{y_thres}{...}
}
\description{
This function plots encapsulated Wilks scores.
}

