% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptlmapper.R
\name{plot_wilks_k}
\alias{plot_wilks_k}
\title{A Function That Plots Encapsulated Wilks Scores for Kantoritch Approach.}
\usage{
plot_wilks_k(ptl_mapping_result, main = "", ...)
}
\arguments{
\item{ptl_mapping_result}{...}

\item{main}{...}
}
\description{
This function plots encapsulated Wilks scores for Kantoritch approach.
}

