% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptlmapper.R
\name{preprocess_data_rqtl}
\alias{preprocess_data_rqtl}
\title{A Function That Preprocesses Genotypes and Phenotypes of Individuals for R Package QTL}
\usage{
preprocess_data_rqtl(genodata_ptl, pheno_hists, kanto_analysis = NULL,
  mmoments_analysis = NULL)
}
\arguments{
\item{genodata_ptl}{The preprocessed genodata. Typically output of `preprocess_genodata` function.}

\item{pheno_hists}{A list of object containing mean and var attribute. Typically outpout of the `build_pheno_hists` function.}

\item{kanto_analysis}{Output of `ptl_scan` function used with the method "kanto".}

\item{mmoments_analysis}{Output of `ptl_scan` function used with the method "mmoments".}
}
\description{
This function preprocesses genotypes and phenotypes of individuals for R Package QTL.
}

