% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptlmapper.R
\name{preprocess_genodata}
\alias{preprocess_genodata}
\title{A Function That Preprocesses Genotype of Individuals.}
\usage{
preprocess_genodata(genodata, bckg)
}
\arguments{
\item{genodata}{A matrix describing the genotype of individuals.}

\item{bckg}{Vector of character describing the individuals as they are describe in genodata.}
}
\description{
This function preprocesses genotype of individuals. 
All the allele that are describe by something else that 1 or will be replaced by NA and ignored by the analysis.
}

