% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptlmapper.R
\name{ptl_mapping}
\alias{ptl_mapping}
\title{A Workflow That Maps PTLs, QTLs, Returns involved Objects, Embed Caching Fetures}
\usage{
ptl_mapping(genodata, cells, bckg, nb_perm = 20, bin_width = NULL,
  nb_dim = 0, errs = 0.05, nb_bin = 100, nb_moments = 4,
  min_prop = 0.1, ptl_mapping_filename = NULL, COMPUTE_KD_MATRIX = TRUE,
  COMPUTE_MM_MATRIX = TRUE, DO_KANTO = TRUE, DO_MMOMENTS = TRUE,
  DO_RQTL = TRUE, CLEAN_OBJECT = FALSE)
}
\arguments{
\item{genodata}{A matrix describing the genotype of individuals.}

\item{cells}{A list of vector of integer (single cell values).}

\item{bckg}{Vector of character describing the individuals as they are describe in genodata.}

\item{nb_perm}{An integer that specifies the number of permuation to do.}

\item{bin_width}{An integer specifying the width of the bin to use.}

\item{nb_dim}{An integer that specifies the number of dimension of the MDS space to explore.}

\item{errs}{A vector of integer (error) that will be used to compute threshold from the permutation test.}

\item{nb_bin}{An integer specifying the number of to use.}

\item{nb_moments}{An integer specifying the number of moments to compute.}

\item{min_prop}{A numeric specifying the minimal proportion of each parental allele under which a marker is discard from the analysis.}

\item{ptl_mapping_filename}{A character string that specifies the file to save the `ptl_mapping` results.}

\item{COMPUTE_KD_MATRIX}{A boolean that specifies if `kd_matrix` needs to be computed.}

\item{COMPUTE_MM_MATRIX}{A boolean that specifies if `mm_matrix` needs to be computed.}

\item{DO_KANTO}{A boolean that specifies if  `ptl_scan` with method="kanto" needs to be performed.}

\item{DO_MMOMENTS}{A boolean that specifies if `ptl_scan` with method="mmoments" needs to be performed.}

\item{DO_RQTL}{A boolean that specifies if `rqtl_launch` needs to be called.}

\item{CLEAN_OBJECT}{A boolean that specifies if `ptl_mapping` results needs to be cleaned.}
}
\description{
This function is a workflow that encapsulated many functions of the R package `ptlmapper`. 
It aggregates parameter values, inputs and outputs of the call to the R package `ptlmapper` function.
The resulting data structure could be cache on the file system. 
It is useful to massivelly save multiple call to `plt_scan` and `rqtl_launch` function, for example in a complex design.
}

