% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptlmapper.R
\name{rqtl_launch}
\alias{rqtl_launch}
\title{A Function That Scans Genome to Detect QTLs (based on R package `qtl`).}
\usage{
rqtl_launch(rqtl_data, nb_perm, errs = 0.05)
}
\arguments{
\item{rqtl_data}{Preprocessed genotypes (typically outpout of preprocess_data_rqtl function).}

\item{nb_perm}{An integer that specifies the number of permuation to do.}

\item{errs}{A vector of integer (error) that will be used to compute threshold from the permutation test.}
}
\description{
This function embed R package `qtl`. It scans genome to detect QTLs using `ptlmapper` data structires.
}

